Modify testing harness so Basic step0 test works
C64 Basic automatically inserts a carriage return (ASCII 13 or \r) at the end of every printed line. You can avoid this by appending a PRINT statement with a semicolon, but this causes the test to hang. Instead, I (temporarily?) put some code into the runtest.py file to check for lines ending with carriage returns and to delete them if found.